+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 ## Unreleased
 
+## 0.2.0
+
+### Changed
+
+- `:rs_la` for serbian (latin) locale renamed to `:rs` (no backward compatibility maintained)
+- All definitions files download are now locked to
+https://github.com/holidays/definitions/tree/v2.3.0 tag
+
+## 0.1.2
+
+### Fixed
+
+- Fixed runtime holidays definitions path
+
+## 0.1.1
+
+### Added
+
+- Dialyzer and better specs
+
 ## 0.1.0
 
 ### Added